adding stuff

+++ srcs/ft_strlen.c
@@ -1,5 +1,6 @@
-unsigned int ft_strlen(const char *s){
-    unsigned int    i;
+#include <stddef.h>
+size_t ft_strlen(const char *s){
+    size_t    i;
 
     i = 0;
     while(s[i])
